==> notebook_85/flo/3. augmentation flo.ipynb <==
import numpy as np # linear algebra
import pandas as pd # data processing, CSV file I/O (e.g. pd.read_csv)
import os
import joblib
import sklearn
# Import Scikit-Learn library for decision tree models
import sklearn 

from sklearn import datasets, linear_model, metrics, tree
from sklearn.utils import resample

from sklearn.tree import DecisionTreeRegressor

from sklearn.ensemble import BaggingRegressor, GradientBoostingRegressor


from sklearn.neighbors import KNeighborsRegressor

        
# Dataset splitting        
from sklearn.model_selection import train_test_split


from matplotlib import pyplot as plt


import os
import joblib

df = pd.read_csv("flo_dataset_scaled.csv")
df

# Input for ML models
input_col = ['in_amount_mmol', 'p_amount_mmol', 'ligand_amount_mmol',
       'first_sol_amount_ml', 'second_sol_amount_ml', 'third_sol_amount_ml',
       'other_1_amount_mmol', 'other_2_amount_mmol', 'total_volume_ml',
       'temp_c', 'time_min', 'x0_chloroindium oxalate', 'x0_indium acetate',
       'x0_indium bromide', 'x0_indium chloride', 'x0_indium iodide',
       'x0_indium myristate', 'x0_indium oxalate', 'x0_indium palmitate',
       'x0_indium trifluoroacetate',
       'x0_indium tris(N,N-diisopropylacetamidinato)',
       'x1_bis(trimethylsilyl)phosphine', 'x1_phosphine gas',
       'x1_phosphorus trichloride', 'x1_sodium phosphide',
       'x1_tris(diethylamino)phosphine', 'x1_tris(dimethylamino)phosphine',
       'x1_tris(trimethylgermyl)phosphine', 'x1_tris(trimethylsilyl)phosphine',
       'x1_white phosphorus', 'x2_None', 'x2_dodecanethiol', 'x2_lauric acid',
       'x2_myristic acid', 'x2_oleic acid', 'x2_palmitic acid',
       'x2_stearic acid', 'x3_4-ethylpyridine', 'x3_None',
       'x3_dimethylformamide', 'x3_dodecylamine', 'x3_mesitylene',
       'x3_octadecene', 'x3_oleylamine', 'x3_trioctylamine',
       'x3_trioctylphosphine', 'x3_trioctylphosphine oxide', 'x4_None',
       'x4_dioctyl ether', 'x4_dioctylamine', 'x4_hexadecylamine',
       'x4_octylamine', 'x4_oleylamine', 'x4_toluene', 'x4_trioctylphosphine',
       'x4_trioctylphosphine oxide', 'x5_None', 'x5_trioctylphosphine',
       'x6_None', 'x6_acetic acid', 'x6_superhydride',
       'x6_tetrabutylammonium myristate', 'x6_zinc acetate', 'x6_zinc bromide',
       'x6_zinc chloride', 'x6_zinc iodide', 'x6_zinc octanoate',
       'x6_zinc oleate', 'x6_zinc stearate', 'x6_zinc undecylenate', 'x7_None',
       'x7_copper bromide', 'x7_oleic acid', 'x7_water', 'x7_zinc iodide']


# Load ML model for predicting absorbance

loaded_rf_ab = joblib.load('model_aug_abs_ExtraTrees.joblib')

# Replace 'None' entries in 'abs_nm' column by predicted values.

a = 0
for index, row in df.iterrows():
    if row['abs_nm'] == 'None':
        X = df.loc[index, input_col].to_numpy()
        df.loc[index, 'abs_nm'] = loaded_rf_ab.predict(X.reshape(1, -1))[0]
        a += 1

        
# Save the dataset where all 'None' values in 'Absorbane max (nm)' column are replaced.   
df.to_csv('dataset_augmented_abs.csv')
df_first_aug = pd.read_csv('dataset_augmented_abs.csv')

#Saves the row indexes to drop for absorbance modelling into a list
total_row_num = len(df_first_aug)
drop_list_emi =[]
for row_i in range(total_row_num):
    if df_first_aug['emission_nm'].values[row_i] == 'None':
        drop_list_emi.append(row_i)
    
len(drop_list_emi)

#Drops the appropriate rows
df_emission_scaled_encoded = df_first_aug.drop(drop_list_emi)
#Saves the data for absorbance modelling to CSV
df_emission_scaled_encoded.to_csv('dataset_scaled_emission.csv')

# Move to **'model flo emission'** notebook to optimize model for augmenting emission

# adding column 'abs_nm' to input
input_col.append('abs_nm')
input_col_2 = input_col.copy()

print(input_col_2)
input_col.remove('abs_nm')

# Now, the dataset only has 'None' values in the 'emission_nm' column

df_emi = pd.read_csv('dataset_augmented_abs.csv') 

# Load ML model for predicting emission

loaded_rf_emi = joblib.load('model_aug_emission_ExtraTrees.joblib')


# # Replace 'None' entries in 'emission_nm' column by predicted values.

a = 0
for index, row in df_emi.iterrows():
    if row['emission_nm'] == 'None':
        X = df_emi.loc[index, input_col_2].to_numpy()
        df_emi.loc[index, 'emission_nm'] = loaded_rf_emi.predict(X.reshape(1, -1))[0]
        a += 1

# # # Save the dataset where all 'None' values are replaced.

# # # Ready to use for other analysis.

df_emi.to_csv('dataset_augmented_abs_emission.csv')


## Aldjust dataset_augmented_abs_emission.csv to dataset_augmented_abs_emission_adjusted.csv

# adding column 'emission_nm' to input
input_col_2.append('emission_nm')
input_col_3 = input_col_2.copy()


print(input_col_3)
input_col_2.remove('emission_nm')

df_second_aug = pd.read_csv('dataset_augmented_abs_emission_adjusted.csv')
#Saves the row indexes to drop for diameter modelling into a list
total_row_num = len(df_second_aug)
drop_list_dia =[]
for row_i in range(total_row_num):
    if df_second_aug['diameter_nm'].values[row_i] == 'None':
        drop_list_dia.append(row_i)
    
len(drop_list_dia)

#Drops the appropriate rows
df_dia_scaled_encoded = df_second_aug.drop(drop_list_dia)
#Saves the data for absorbance modelling to CSV
df_dia_scaled_encoded.to_csv('dataset_scaled_diameter.csv')

# Move to **'model hao diameter'** notebook to optimize model for augmenting emission.


# Load ML model for predicting diameters

df_dia = pd.read_csv('dataset_augmented_abs_emission_adjusted.csv') 

loaded_rf_dia = joblib.load('model_aug_diameter_Bagging.joblib')

# Replace 'None' entries in 'diameter_nm' column by predicted values.

a = 0
for index, row in df_dia.iterrows():
    if row['diameter_nm'] == 'None':
        X = df_dia.loc[index, input_col_3].to_numpy()
        df_dia.loc[index, 'diameter_nm'] = loaded_rf_dia.predict(X.reshape(1, -1))[0]
        a += 1

        
# Save the dataset where all 'None' values in 'diameter_nm' column are replaced.   
df_dia.to_csv('flo_dataset_augmented.csv')
